update deepdao models
