+++ daochem/database/models/daos.py
@@ -2,7 +2,6 @@ from django.db import models
 
 from daochem.database.models.blockchain import BlockchainAddress
 from daochem.database.models.twitter import TwitterAccount
-from utils.strings import clean_dao_name
 
 _STR_KWARGS = {'max_length': 200, 'null': True}
 
@@ -13,40 +12,6 @@ SURVEY_QUESTIONS = {
     'q4': "do you feel wanted and/or needed by the DAO?", 
     'q5': "does contributing to this DAO bring you a sense of fulfillment?", 
 }
-
-
-class DeepdaoEntry(models.Model):
-    organization_name = models.CharField(max_length=50, default="")
-    description = models.CharField(max_length=50, null=True)
-    url = models.URLField(**_STR_KWARGS)
-
-    @property
-    def name_cleaned(self):
-        return clean_dao_name(self.organization_name)
-
-    class Meta:
-        db_table = "deepdao_daos"
-
-    def __str__(self):
-        return self.id
-
-
-
-class DeepdaoAddress(models.Model):
-    id = models.CharField(primary_key=True, max_length=200, default="")
-    type = models.CharField(max_length=50, null=True)
-    address = models.ForeignKey(
-        BlockchainAddress,
-        on_delete=models.CASCADE,
-        null=True,
-        related_name='deepdao_info'
-    )
-
-    class Meta:
-        db_table = "deepdao_addresses"
-
-    def __str__(self):
-        return self.id
 
 
 class Dao(models.Model):
@@ -64,10 +29,6 @@ class Dao(models.Model):
     boardroom = models.URLField(**_STR_KWARGS)
     governance_addresses = models.ManyToManyField(
         BlockchainAddress,
-        related_name='belongs_to_dao'
-    )
-    deepdao_addresses = models.ManyToManyField(
-        DeepdaoAddress,
         related_name='belongs_to_dao'
     )
